requirements.txt
streamlit
pandas
XlsxWriter
python-docx